refactor(sqs): Consolidate `MessageAttributes` conversion
Put from / to `MessageAttributes` conversion into a dedicated file for a
better overview.
